Z buffer on offscreen is working, but tests are still failing because drawing doesn't work

+++ Model/GPUCalculationEngine.h
@@ -125,7 +125,7 @@ namespace hdsim {
       	/**
           * Rendered depth buffer (from the graphic card)
           */
-	      double *renderedDepth_;
+	      GLfloat *renderedDepth_;
 
 	};
